Replace org.apache.regexp package with java.util.regex
The former is ancient, unmaintained, non-standard, and doesn't really
support Perl regular expressions (which may be used by PHPTs), and
appears to be grossly buggy.  Therefore we replace it with the Java SE
regular expression package, which is already used elsewhere anyway.

We also simplify the conversion of EXPECTF strings to regexps by doing
simple string replacements instead of regexp replacements, and make the
matching more strict by actually anchoring the pattern to avoid false
negatives.

Finally, we remove the regex debug stuff, since none of it is supported
by java.util.regex; if necessary, we will have to use a debugger.

This fixes #79.